fix(hooks): break hooks-types <-> hooks-schemas import cycle
The schema-derived-types change introduced a two-way import (hooks-schemas
needed HookName/HookDefinition from hooks-types; hooks-types re-exported the
derived payload types from hooks-schemas), tripping the structural gate's
cycle check. HookName, HookDefinition, and HookRegistry now live in
hooks-schemas.ts next to the registry they key; hooks-types re-exports them,
so the public import surface (and the deep './hooks-types.js' imports across
src and tests) are unchanged. Dependency is now strictly one-directional:
hooks-types -> hooks-schemas.
